Moves requests integration tests to a separate folder
