+++ test/integration/requests/reports_spec.js
@@ -1,4 +1,4 @@
-import app from '../../src/app';
+import app from '../../../src/app';
 
 describe('POST /reports', () => {
   context('invalid', () => {
